requirements.txt
numpy
fastapi
uvicorn[standard]
grpcio
grpcio-tools
faker
openai-whisper
kafka-python
transformers
spacy